requirements.txt
annotated-types==0.6.0
anyio==4.2.0
click==8.1.7
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.109.0
h11==0.14.0
idna==3.6
pip==23.3.2
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.5.3
pydantic-core==2.14.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
starlette==0.35.1
typing-extensions==4.9.0
uvicorn==0.25.0
wheel==0.42.0
